buffetiser/backend/requirements.txt
django
django-cors-headers
psycopg2-binary
djangorestframework
django-rest-swagger
requests
beautifulsoup4
aiohttp
gunicorn
schedule
pytz
